Cardano.Slotting.EpochInfo.API

Documentation

data EpochInfo m

Information about epochs

Different epochs may have different sizes and different slot lengths. This information is encapsulated by EpochInfo. It is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself. It ultimately requires acess to state, and so either uses the monad for that or uses the monad to reify failure due to cached state information being too stale for the current query.

Constructors

EpochInfo

Fields

epochInfoSize_ ∷ HasCallStack ⇒ EpochNo → m EpochSize

Return the size of the given epoch as a number of slots

Note that the number of slots does not bound the number of blocks, since the EBB and a regular block share a slot number.

epochInfoFirst_ ∷ HasCallStack ⇒ EpochNo → m SlotNo

First slot in the specified epoch

See also epochInfoRange

epochInfoEpoch_ ∷ HasCallStack ⇒ SlotNo → m EpochNo

Epoch containing the given slot

We should have the property that

s `inRange` epochInfoRange (epochInfoEpoch s)

epochInfoSlotToRelativeTime_ ∷ HasCallStack ⇒ SlotNo → m RelativeTime

The RelativeTime of the start of the given slot

This calculation depends on the varying slot lengths of the relevant epochs.

See also epochInfoSlotToUTCTime.

epochInfoSlotLength_ ∷ HasCallStack ⇒ SlotNo → m SlotLength

Return the length of the specified slot.

epochInfoSize ∷ HasCallStack ⇒ EpochInfo m → EpochNo → m EpochSize

epochInfoFirst ∷ HasCallStack ⇒ EpochInfo m → EpochNo → m SlotNo

epochInfoEpoch ∷ HasCallStack ⇒ EpochInfo m → SlotNo → m EpochNo

epochInfoRange ∷ Monad m ⇒ EpochInfo m → EpochNo → m (SlotNo, SlotNo)

epochInfoSlotToRelativeTime ∷ HasCallStack ⇒ EpochInfo m → SlotNo → m RelativeTime

epochInfoSlotToUTCTime ∷ (HasCallStack, Monad m) ⇒ EpochInfo m → SystemStart → SlotNo → m UTCTime

The start of the given slot

epochInfoSlotLength ∷ HasCallStack ⇒ EpochInfo m → SlotNo → m SlotLength

Utility

hoistEpochInfo ∷ (∀ a. m a → n a) → EpochInfo m → EpochInfo n

generalizeEpochInfo ∷ Monad m ⇒ EpochInfo Identity → EpochInfo m