Cardano.Slotting.EpochInfo.Impl

Description

For use in trivial cases, such as in mocks, tests, etc.

Documentation

fixedEpochInfo :: Monad m => EpochSize -> SlotLength -> EpochInfo m

The EpochInfo induced by assuming the epoch size and slot length are fixed for the entire system lifetime

Shortcuts

fixedEpochInfoEpoch :: EpochSize -> SlotNo -> EpochNo

The pure computation underlying epochInfoEpoch applied to fixedEpochInfo

You don't need a SlotLength for this.

fixedEpochInfoFirst :: EpochSize -> EpochNo -> SlotNo

The pure computation underlying epochInfoFirst applied to fixedEpochInfo

You don't need a SlotLength for this.