Cardano.Slotting.SlotBounded

Description

Intended for qualified import

Bounds

data Bounds

Constructors

II
Both bounds are inclusive

IX
Lower bound is inclusive, upper bound is exclusive

class InBounds (bounds ∷ Bounds) where

Methods

inBounds ∷ proxy bounds → WithOrigin SlotNo → (WithOrigin SlotNo, SlotNo) → Bool

Slot-bounded values

data SlotBounded (bounds ∷ Bounds) a

An item bounded to be valid within particular slots

Constructors

SlotBounded

Fields

sbLower ∷ !(WithOrigin SlotNo)

sbUpper ∷ !SlotNo

sbContent ∷ !a

bounds ∷ SlotBounded bounds a → (WithOrigin SlotNo, SlotNo)

bounded ∷ WithOrigin SlotNo → SlotNo → a → SlotBounded bounds a

Construct a slot bounded item.

We choose not to validate that the slot bounds are reasonable here.

maximal ∷ a → SlotBounded bounds a

at ∷ InBounds bounds ⇒ SlotBounded bounds a → WithOrigin SlotNo → Maybe a

contains ∷ ∀ bounds a. InBounds bounds ⇒ SlotBounded bounds a → WithOrigin SlotNo → Bool