Cardano.Slotting.EpochInfo.API

Documentation

data EpochInfo m

Information about epochs

Epochs may have different sizes at different times during the lifetime of the blockchain. This information is encapsulated by EpochInfo; it is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself, and hence requires access to state.

The other functions provide some derived information from epoch sizes. In the default implementation all of these functions query and update an internal cache maintaining cumulative epoch sizes; for that reason, all of these functions live in a monad m.

Constructors

EpochInfo

Fields

epochInfoSize ∷ EpochNo → m EpochSize

Return the size of the given epoch as a number of slots

Note that the number of slots does not bound the number of blocks, since the EBB and a regular block share a slot number.

epochInfoFirst ∷ EpochNo → m SlotNo

First slot in the specified epoch

See also epochInfoRange

epochInfoEpoch ∷ SlotNo → m EpochNo

Epoch containing the given slot

We should have the property that

s `inRange` epochInfoRange (epochInfoEpoch s)

epochInfoRange ∷ Monad m ⇒ EpochInfo m → EpochNo → m (SlotNo, SlotNo)

Utility

hoistEpochInfo ∷ (∀ a. m a → n a) → EpochInfo m → EpochInfo n

generalizeEpochInfo ∷ Monad m ⇒ EpochInfo Identity → EpochInfo m