Cardano.Slotting.EpochInfo.Extend

Documentation

unsafeLinearExtendEpochInfo ∷ Monad m ⇒ SlotNo → EpochInfo m → EpochInfo m

Given a basis point, use it and its slot length to impute a linear relationship between slots and time in order to extend an EpochInfo to infinity.

The returned EpochInfo may still fail (according to the semantics of the specified monad) if any of the underlying operations fail. For example, if we cannot translate the basis point.